+++ lib/src/services/log_service/log_level.dart
@@ -1,20 +1,12 @@
-enum LogLevel { debug, verbose, info, warning, error, critical }
+enum LogLevel {
+  debug(0),
+  verbose(1),
+  info(2),
+  warning(3),
+  error(4),
+  critical(5);
 
-extension SeverityLogLevelExtension on LogLevel {
-  int toSeverity() {
-    switch (this) {
-      case LogLevel.debug:
-        return 0;
-      case LogLevel.verbose:
-        return 1;
-      case LogLevel.info:
-        return 2;
-      case LogLevel.warning:
-        return 3;
-      case LogLevel.error:
-        return 4;
-      case LogLevel.critical:
-        return 5;
-    }
-  }
+  final int level;
+
+  const LogLevel(this.level);
 }